added side-bar login

+++ client/src/pages/SignUp.js
@@ -15,37 +15,42 @@ function SignUp({userUI,adminUI}) {
         e.preventDefault();
       try {
         if(password===confirmPassword){
-            try {
-                const url=`http://localhost:5000/api/register`;
-                const response= await fetch(url,{
-                    method:"POST",
-                    body:JSON.stringify({
-                        pic:1,
-                        username,
-                        email,
-                        password:confirmPassword
-                    }),
-                    headers:{
-                        "content-type":"application/json"
+            if(window.confirm("MAKE SURE YOU'VE READ OUR TERMS AND CONDITIONS BEFORE SIGNING UP!!")){
+                try {
+                    const url=`http://localhost:5000/api/register`;
+                    const response= await fetch(url,{
+                        method:"POST",
+                        body:JSON.stringify({
+                            pic:1,
+                            username,
+                            email,
+                            password:confirmPassword
+                        }),
+                        headers:{
+                            "content-type":"application/json"
+                        }
+                    })
+                    const parseRes=await response.json();
+                    if(parseRes.error){
+                        toast.error(parseRes.error);
+                        document.querySelector('form').reset()
+                    }else{
+                        toast.success(parseRes.msg);
+                        localStorage.setItem('userID',parseRes._id);
+                        localStorage.setItem('pic',parseRes.pic);
+                        localStorage.setItem('username',parseRes.username);
+                        localStorage.setItem('email',parseRes.email);
+                        sessionStorage.setItem('userToken',parseRes.token);
+                        navigate('/');
+                        setTimeout(()=>{
+                            window.location.reload()
+                        },200)
                     }
-                })
-                const parseRes=await response.json();
-                if(parseRes.error){
-                    toast.error(parseRes.error)
-                }else{
-                    toast.success(parseRes.msg);
-                    localStorage.setItem('userID',parseRes._id);
-                    localStorage.setItem('pic',parseRes.pic);
-                    localStorage.setItem('username',parseRes.username);
-                    localStorage.setItem('email',parseRes.email);
-                    sessionStorage.setItem('userToken',parseRes.token);
-                    navigate('/');
-                    setTimeout(()=>{
-                        window.location.reload()
-                    },200)
+                } catch (error) {
+                    toast.error(error.message);
                 }
-            } catch (error) {
-                toast.error(error.message);
+            }else{
+                navigate('/guide')
             }
            }else{
             toast.error("Password doesn't match ☠!")
@@ -56,7 +61,7 @@ function SignUp({userUI,adminUI}) {
     }
     return (
         <>
-        <Navbar userUI={userUI} adminUI={adminUI}/>
+        <Navbar/>
             <div className='sign-in '>
                 <form onSubmit={handleRegister}>
                     <label>Username</label>
@@ -68,7 +73,7 @@ function SignUp({userUI,adminUI}) {
                     <label>Confirm Password</label>
                     <input type='password' onChange={e=>setConfirmPasword(e.target.value)} required/>
                     <label>Upload your avatar</label>
-                    <input type='file' onChange={e=>setFile(e.target.value)} required/>
+                    <input type='file' accept='image/x-png, image/jpg, image/jpeg' onChange={e=>setFile(e.target.value)} required/>
                     <div className='link-btn'>
                         <button>Submit</button> <Link to='/login'>I have an account?</Link>
                     </div>
